Update for Swift 3
Sweeping changes, done with the help of the automatic migrator in xCode 8.
It looks like a lot, but it's all pretty straightforward.
The biggest question is whether to use the new URL and Data value types instead of NSURL and NSData. I think it's probably best to commit to URL and Data right away.
All tests pass!